Initial auth move to modals
