+++ bartop-ui/src/components/App/App.js
@@ -8,16 +8,18 @@ import { Router, Route, Switch, Redirect } from 'react-router-dom';
 import { injectGlobal } from 'styled-components';
 import ApolloClient from 'apollo-boost';
 import { ApolloProvider } from 'react-apollo';
+import ReactModal from 'react-modal';
 
 import config from '../../config';
 import callApiMiddleware from '../../middleware/call-api';
 import rootReducer from '../../ducks';
-import auth from '../../singletons/authentication';
+import { actions as authActions } from '../../ducks/authentication/authentication';
 import history from '../../singletons/history';
 import Landing from '../Landing/Landing';
-import Auth from '../Auth/Auth';
+import Callback from '../Callback/Callback';
 import NotFound from '../NotFound/NotFound';
 import { colors } from '../styleUtils';
+import ModalRoot from '../ModalRoot/ModalRoot';
 
 // persist all of our auth & user state to local storage
 const localStorageEnhancer = persistState(['authentication', 'user'], {
@@ -47,42 +49,50 @@ const client = new ApolloClient({
   }
 });
 
-export default class App extends React.Component {
-  render() {
-    // styled components helper for adding styles to global dom elements like body
-    injectGlobal`
-      html,
-      body,
-      #root {
-        width: 100%;
-        height: 100%;
-        margin: 0;
-        padding: 0;
-        font-family: 'Raleway', sans-serif;
-        color: ${colors.darkest};
-        /* don't use default styling on links */
-        a {
-          color:inherit;
-          text-decoration: none;
-        }
-      }
-    `;
-    return (
-      <ReduxProvider store={store}>
-        <ApolloProvider client={client}>
-          <Router history={history}>
-            <Switch>
-              <Redirect exact from="/" to="/landing" />
-              <Route path="/landing" component={Landing} />
-              <Route
-                path="/auth"
-                render={({ match }) => <Auth match={match} auth={auth} />}
-              />
-              <Route component={NotFound} />
-            </Switch>
-          </Router>
-        </ApolloProvider>
-      </ReduxProvider>
-    );
+// styled components helper for adding styles to global dom elements like body
+injectGlobal`
+html,
+body,
+#root {
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  padding: 0;
+  font-family: 'Raleway', sans-serif;
+  color: ${colors.darkest};
+  /* don't use default styling on links */
+  a {
+    color:inherit;
+    text-decoration: none;
   }
 }
+`;
+
+ReactModal.setAppElement('#root');
+
+const App = () => (
+  <ReduxProvider store={store}>
+    <ApolloProvider client={client}>
+      <Router history={history}>
+        <React.Fragment>
+          <Switch>
+            <Redirect exact from="/" to="/landing" />
+            <Route path="/landing" component={Landing} />
+            <Route
+              exact
+              path="/callback"
+              render={() => {
+                store.dispatch(authActions.handleAuthentication());
+                return <Callback />;
+              }}
+            />
+            <Route component={NotFound} />
+          </Switch>
+          <ModalRoot />
+        </React.Fragment>
+      </Router>
+    </ApolloProvider>
+  </ReduxProvider>
+);
+
+export default App;
