+++ bartop-ui/src/ducks/index.js
@@ -2,8 +2,10 @@ import { combineReducers } from 'redux';
 
 import authenticationReducer from './authentication/authentication';
 import userReducer from './user/user';
+import modalsReducer from './modals/modals';
 
 export default combineReducers({
   authentication: authenticationReducer,
-  user: userReducer
+  user: userReducer,
+  modals: modalsReducer
 });
