+++ bartop-ui/src/ducks/modals/modals.js
@@ -1,3 +1,10 @@
+export const MODAL_TYPES = {
+  DEFAULT_MODAL: 'DEFAULT_MODAL',
+  LOGIN_MODAL: 'LOGIN_MODAL',
+  VERIFY_CODE_MODAL: 'VERIFY_CODE_MODAL',
+  LOGIN_FAILURE_MODAL: 'LOGIN_FAILURE_MODAL'
+};
+
 export const types = {
   SHOW_MODAL: 'MODALS/SHOW_MODAL',
   HIDE_MODAL: 'MODALS/HIDE_MODAL'
