+++ bartop-ui/src/singletons/authentication.js
@@ -8,7 +8,7 @@ export const loginsCountClaim = `${config.auth0.claimNamespace}/loginsCount`;
 export default new auth0.WebAuth({
   domain: config.auth0.domain,
   clientID: config.auth0.clientId,
-  redirectUri: `${config.url}/auth/callback`,
+  redirectUri: `${config.url}/callback`,
   audience: `${config.auth0.apiAudience}`,
   responseType: 'token id_token',
   // check out https://openid.net/specs/openid-connect-core-1_0.html#StandardClaims
